src/main/java/core/CampaignController.java
package core;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;
import service.CampaignService;

import java.util.List;

/**
 * Campaign controller.
 */
@RestController
@RequestMapping( value = "/campaign" )
public class CampaignController {

    @Autowired
    private CampaignService campaignService;

    @RequestMapping(method = RequestMethod.GET)
    public List<Campaign> findAll() {
        return campaignService.findAll();
    }

    @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
    @ResponseStatus(HttpStatus.CREATED)
    public Campaign createCampaign(@RequestBody Campaign newCampaign) {
        return campaignService.create(newCampaign);
    }
}